feat(): create trip handler

+++ src/migrations/1662815032120-Trip.ts
@@ -5,7 +5,7 @@ export class Trip1662815032120 implements MigrationInterface {
 
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(
-      `CREATE TABLE "trips" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "start_address" character varying NOT NULL, "destination_address" character varying NOT NULL, "distance" integer NOT NULL, "price" integer NOT NULL, "currency" character varying NOT NULL, "delivery_date" TIMESTAMP WITH TIME ZONE NOT NULL, CONSTRAINT "PK_f71c231dee9c05a9522f9e840f5" PRIMARY KEY ("id"))`,
+      `CREATE TABLE "trips" ("id" uuid NOT NULL DEFAULT uuid_generate_v4(), "start_address" character varying NOT NULL, "destination_address" character varying NOT NULL, "distance" integer NOT NULL, "price" numeric(8,2) NOT NULL, "currency" character varying NOT NULL, "delivery_date" TIMESTAMP WITH TIME ZONE NOT NULL, CONSTRAINT "PK_f71c231dee9c05a9522f9e840f5" PRIMARY KEY ("id"))`,
     );
   }
